feature: Add datatype BinUUID for uuid as binary storage in DB
This adds a new datatype BinUUID, similar to datatypes.UUID but
with a major difference being that BinUUID stores the uuid in
the database as a binary (byte) array instead of a string.
Developers may use either datatypes.UUID or datatypes.BinUUID
as per their preference, either storing uuid in the database
as a string or as a binary (byte) array respectively.